pyproject.toml
[project]
name = "atoms-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "boto3>=1.42.35",
    "fastapi>=0.128.0",
    "google-auth>=2.48.0",
    "google-cloud-aiplatform>=1.134.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
